More channel related conversions
