+++ components/builder-api/src/server/resources/jobs.rs
@@ -34,6 +34,7 @@ use server::error::{Error, Result};
 use server::framework::headers;
 use server::framework::middleware::route_message;
 use server::helpers::{self, Target};
+use server::resources::channels::channels_for_package_ident;
 use server::AppState;
 
 #[derive(Clone, Serialize, Deserialize, Debug)]
@@ -407,7 +408,7 @@ fn do_get_job(req: &HttpRequest<AppState>, job_id: u64) -> Result<String> {
             authorize_session(req, Some(&job.get_project().get_origin_name()))?;
 
             if job.get_package_ident().fully_qualified() {
-                let channels = helpers::channels_for_package_ident(req, job.get_package_ident());
+                let channels = channels_for_package_ident(req, job.get_package_ident())?;
                 let platforms = helpers::platforms_for_package_ident(req, job.get_package_ident());
                 let mut job_json = serde_json::to_value(job).unwrap();
