+++ components/builder-api/src/server/resources/projects.rs
@@ -37,6 +37,7 @@ use server::error::Error;
 use server::framework::headers;
 use server::framework::middleware::route_message;
 use server::helpers::{self, Pagination};
+use server::resources::channels::channels_for_package_ident;
 use server::AppState;
 
 #[derive(Clone, Serialize, Deserialize)]
@@ -440,7 +441,10 @@ fn get_jobs((pagination, req): (Query<Pagination>, HttpRequest<AppState>)) -> Ht
                 .map(|job| {
                     if job.get_state() == JobState::Complete {
                         let channels =
-                            helpers::channels_for_package_ident(&req, &job.get_package_ident());
+                            match channels_for_package_ident(&req, &job.get_package_ident()) {
+                                Ok(channels) => channels,
+                                Err(_) => None,
+                            };
                         let platforms =
                             helpers::platforms_for_package_ident(&req, &job.get_package_ident());
                         let mut job_json = serde_json::to_value(job).unwrap();
